backend/requirements.txt
Authlib==1.3.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.0.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
zipp==3.23.0
Flask-SQLAlchemy
Flask-Migrate
psycopg[binary]
